Changed so animals dont change starvation rate in constructor. Started tweaking values.

+++ src/main/java/org/primal/entity/Carnivore.java
@@ -16,7 +16,7 @@ public abstract class Carnivore extends Animal {
      * @param thirst = thirst level
      */
 
-    public Carnivore(float x, float y, Map map, float health, float stamina, float fullness, float thirst) {
+    public Carnivore(float x, float y, Map map, double health, double stamina, double fullness, double thirst) {
         super(x, y, map, health, fullness, stamina, thirst);
     }
